requirements.txt

uvicorn[standard]
fastapi 
pydantic
groq
python-dotenv
sqlalchemy
psycopg2-binary
pymupdf
python-multipart
passlib[bcrypt]
pydantic[email]
python-jose
python-jose[cryptography] 
passlib
jwt
